src/421B/test_cdf_421B.py
import unittest
from unittest.mock import patch

from cdf_421B import CodeforcesTask421BSolution


class TestCDF421B(unittest.TestCase):
    def test_421B_acceptance_1(self):
        mock_input = ['AHA']
        expected = 'YES'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask421BSolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)

    def test_421B_acceptance_2(self):
        mock_input = ['Z']
        expected = 'NO'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask421BSolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)

    def test_421B_acceptance_3(self):
        mock_input = ['XO']
        expected = 'NO'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask421BSolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)


if __name__ == "__main__":
    unittest.main()
